tools: specreader: refactor as a lib instead of a main package
