+++ toolkit/tools/pkg/specreader/config.go
@@ -1,0 +1,13 @@
+package specreader
+
+type Config struct {
+	SpecsDir  string
+	Output    string
+	Workers   int
+	BuildDir  string
+	SrpmsDir  string
+	RpmsDir   string
+	DistTag   string
+	WorkerTar string
+	RunCheck  bool
+}
